+++ RELEASENOTES.md
@@ -20,6 +20,10 @@
     *   Fix a bug when [depending on ExoPlayer locally](README.md#locally) with
         a relative path
         ([#9403](https://github.com/google/ExoPlayer/issues/9403)).
+    *   Fix bug in `DefaultDrmSessionManager` which prevented
+        `MediaSourceFactory` instances from being re-used by `ExoPlayer`
+        instances with non-overlapping lifecycles
+        ([#9099](https://github.com/google/ExoPlayer/issues/9099)).
 *   Extractors:
     *   Support TS packets without PTS flag
         ([#9294](https://github.com/google/ExoPlayer/issues/9294)).